+++ immutable-arrays/core/src/main/kotlin/com/danrusu/pods4k/immutableArrays/ImmutableShortArray.kt
@@ -357,6 +357,17 @@ public value class ImmutableShortArray @PublishedApi internal constructor(
     }
 
     /**
+     * Returns an immutable array containing all the elements expect the last n elements.
+     *
+     * @throws IllegalArgumentException if [n] is negative.
+     */
+    public fun dropLast(n: Int): ImmutableShortArray {
+        require(n >= 0) { "Requested element count $n is less than zero." }
+
+        return take((size - n).coerceAtLeast(0))
+    }
+
+    /**
      * Returns an immutable array containing only the elements matching the given [predicate].
      */
     public inline fun filter(predicate: (element: Short) -> Boolean): ImmutableShortArray {